backend/segmentation-service/requirements.txt
flask
pandas
numpy
kafka-python
psycopg2-binary
pytest
python-dotenv
sqlalchemy